simplifiing the if trees into multiple factors of conditions
The code would be a massive unreadable mess if I continued branching into all the different instances of play that could occur. So instead I'm gathering data points for the pc decision on which card to play. then writing if statements in line with "and"s in the conditionals that match. That way all branches are instead in a same indented line.